requirements.txt
langchain==0.3.20
openai==1.66.3
langchain-openai==0.3.8
azure-search-documents==11.5.2
azure-identity==1.21.0
langchain-community==0.3.19
python-dotenv==1.0.1
beautifulsoup4==4.12.3
autogen-agentchat~=0.2
autogen==0.8.1
